api/requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
openpyxl>=3.0.0
python-dotenv>=1.0.0
supabase>=1.0.0
openai>=1.0.0
